Add frame rate debug data source
