+++ HRCounter/HRCounter.csproj
@@ -150,11 +150,12 @@
     <Compile Include="AssetBundleManager.cs" />
     <Compile Include="Data\BPM.cs" />
     <Compile Include="Data\DataSources\DataSource.cs" />
+    <Compile Include="Data\DataSources\DebugSource\FrameRateHR.cs" />
+    <Compile Include="Data\DataSources\DebugSource\RandomHR.cs" />
+    <Compile Include="Data\DataSources\DebugSource\SweepHR.cs" />
     <Compile Include="Data\DataSources\FitbitHRtoWS.cs" />
     <Compile Include="Data\DataSources\HRProxy.cs" />
     <Compile Include="Data\DataSources\Pulsoid.cs" />
-    <Compile Include="Data\DataSources\RandomHR.cs" />
-    <Compile Include="Data\DataSources\SweepHR.cs" />
     <Compile Include="Data\DataSources\WebRequest.cs" />
     <Compile Include="Data\DataSources\YURApp.cs" />
     <Compile Include="Data\DataSources\YURMod.cs" />
